Add quark architecture
..for the Intel Quark X1000 chip on the Galileo boards.
The chip is "Pentium Compatible" with no mmx or sse extensions,
but has a bug:

71538 - Linux segfault when using lock prefix instruction under
specific circumstances

To work around this everything needs to be assembled with
-momit-lock-prefix=yes, but this means the code won't work
properly on multicore or multithreaded CPUs.

For this reason use quark and i586-quark-linux-gnu rather
than the generic i586 and i586-unknown-linux-gnu for
architecture short name and host triplet.
This way it doesn't clash with a hypothetical future
generic Pentium port.